+++ Raptor/sources/Route.cpp
@@ -1,7 +1,7 @@
 
 #include "Route.h"
 
-void Route::set(int p_id, int p_nb_stops, int p_nb_trips){
+void Route::set(long long p_id, long long p_nb_stops, long long p_nb_trips){
   id = p_id;
   nb_stops = p_nb_stops;
   nb_trips = p_nb_trips;
